requirements.txt
Flask==2.3.2
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1        # only if you really send mail
Flask-MySQL==1.6.0       # only if you use MySQL
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2         # only if you use WTForms
boto3==1.40.32
psycopg2-binary==2.9.10  # Postgres driver
Pillow==10.4.0
python-dotenv==1.0.1
gunicorn==21.2.0

# Missing dependencies for face recognition
numpy>=1.24.0
face-recognition>=1.3.0
dlib>=19.24.0
 